refactor(tip403): migrate TIP403Registry from sol! to #[abi] macro
- Add abi.rs with Interface trait, Error/Event enums, PolicyType with Storable
- Update mod.rs to use new Interface trait with individual params
- Update dispatch.rs for new API
- Update error.rs with new error constructors
- Fix ripple effects in tip20, stablecoin_dex, transaction-pool
- Update benchmarks for new API

Amp-Thread-ID: https://ampcode.com/threads/T-019c2a14-b29e-70c9-9473-c4aa9cdd3c9a

+++ crates/transaction-pool/src/validator.rs
@@ -1418,7 +1418,7 @@ mod tests {
 
         // Add TIP403Registry with blacklist policy containing fee_payer
         let policy_data = PolicyData {
-            policy_type: ITIP403Registry::PolicyType::BLACKLIST as u8,
+            policy_type: ITIP403Registry::PolicyType::BLACKLIST,
             admin: Address::ZERO,
         };
         let policy_data_slot = TIP403Registry::new().policy_records[policy_id]
